+++ consensus/enclave/impl/src/constant_time_token_map.rs
@@ -400,7 +400,7 @@ mod tests {
             .collect()
     }
 
-    fn token_id_set(ids: &[u32]) -> BTreeSet<TokenId> {
+    fn token_id_set(ids: &[u64]) -> BTreeSet<TokenId> {
         ids.iter().map(|id| TokenId::from(*id)).collect()
     }
