+++ mobilecoind/src/utxo_store.rs
@@ -58,8 +58,8 @@ pub struct UnspentTxOut {
     pub attempted_spend_tombstone: u64,
 
     /// The token id of this TxOut
-    #[prost(uint32, tag = "7")]
-    pub token_id: u32,
+    #[prost(uint64, tag = "7")]
+    pub token_id: u64,
 }
 
 /// Type used as the key in the utxo_id_to_utxo  database.
